feat(ai_chat): 全链路修复 — listen start/VAD/播放流畅/全量字体/按键音
AI 对话从"卡死无响应"到完整闭环：
- 协议：连接后发 SendStartListening（服务器才处理音频，缺失=音频全丢弃）
- 简易 VAD：语音后静音 900ms 自动停止（服务器不做判停）
- Opus 预分配内存归还复用（destroy 释放后碎片化致第二轮无声）
- 播放连续流：追加式不打断、DMA 缓冲 240ms、解码失败补静音帧、400ms jitter
- 全量中文字体 GB2312 6763 字（一级+二级串联 fallback，RLE 压缩）
- AI 对话禁用按键提示音（20ms 高音=滴滴声）
- 24kHz 直采对齐服务器，移除重采样

+++ main/framework/ai_chat/ai_resample.cc
@@ -23,11 +23,14 @@ size_t ai_resample_linear(const int16_t* in, size_t in_len,
         return n;
     }
 
-    const double step = (double)in_rate / (double)out_rate;
-    size_t out_idx    = 0;
-    for (double phase = 0.0; phase < in_len - 1 && out_idx < out_cap; phase += step, ++out_idx) {
+    // 注意用 float 而非 double：xtensa 无 double FPU，软浮点库（__muldf3 等）
+    // 栈需求巨大，在录音任务中会栈溢出写穿破坏堆。float 有硬件 FPU，精度足够
+    const float step = (float)in_rate / (float)out_rate;
+    size_t out_idx   = 0;
+    for (float phase = 0.0f; phase < (float)(in_len - 1) && out_idx < out_cap;
+         phase += step, ++out_idx) {
         size_t i0    = static_cast<size_t>(phase);
-        float frac   = static_cast<float>(phase - i0);
+        float frac   = phase - (float)i0;
         float sample = in[i0] * (1.0f - frac) + in[i0 + 1] * frac;
         out[out_idx] = static_cast<int16_t>(sample);
     }
